+++ behaviors/croquet/halfBodyAvatar.js
@@ -3,11 +3,17 @@ class AvatarActor {
         this._cardData.animationClipIndex = 9;
         this.say("animationStateChanged");
         this.listen("poseAvatarRequest", "poseAvatar");
+        this.listen("setAvatarData", "resetPose");
     }
 
     poseAvatar(data) {
         this.lastPose = data;
         this.say("avatarPosed", data);
+    }
+
+    resetPose() {
+        this.lastPose = {type: "move", coordinates: [0, 1, -100], pointing: false};
+        this.say("avatarPosed", this.lastPose);
     }
 }
 
@@ -47,39 +53,18 @@ class AvatarPawn {
 
         this.addLastResponder("keyUp", {ctrlKey: true}, this);
         this.addEventListener("keyUp", this.keyUp);
-    }
-
-    startMotion(dx, dy) {
-        this.spin = Microverse.q_identity();
-        this.velocity = Microverse.v3_zero();
-        this.say("startFalling");
-        if (dx || dy) this.updateMotion(dx, dy);
-    }
-
-    endMotion(_dx, _dy) {
-        this.activeMMotion = false;
-        this.spin = Microverse.q_identity();
-        this.velocity = Microverse.v3_zero();
-    }
-
-    updateMotion(dx, dy) {
-        const JOYSTICK_V = 0.000030;
-        const MAX_V = 0.015;
-        const MAX_SPIN = 0.0004;
-
-        let v = dy * JOYSTICK_V;
-        v = Math.min(Math.max(v, -MAX_V), MAX_V);
-
-        const yaw = dx * (this.isMobile ? -2.5 * MAX_SPIN : -MAX_SPIN);
-        this.spin = Microverse.q_euler(0, yaw ,0);
-        this.velocity = [0, 0, v];
-        this.maybeLeavePresentation();
+        this.addUpdateRequest(["HalfBodyAvatarEventHandler$AvatarPawn", "maybeMove"]);
     }
 
     handlingEvent(type, target, event) {
         if (type.startsWith("pointer")) {
             const render = this.service("ThreeRenderManager");
-            const rc = this.pointerRaycast(event.xy, render.threeLayerUnion('pointer', "walk"));
+            let rc;
+            if (type === "pointerDown") {
+                rc = this.pointerRaycast(event.xy, render.threeLayerUnion('pointer', "walk"));
+            } else {
+                rc = this.pointerRaycast(event.xy, render.threeLayerUnion("walk"));
+            }
             let p3e = this.pointerEvent(rc, event);
             this.move(type, p3e.xyz);
         }
@@ -104,7 +89,7 @@ class AvatarPawn {
             console.log("ready player me avatar found");
         }
 
-        this.moveRightHand([0, 1, -100]);
+        this.addFoot();
 
         if (this.actor.lastPose) {
             this.avatarPosed(this.actor.lastPose);
@@ -124,19 +109,43 @@ class AvatarPawn {
         */
     }
 
+    addFoot() {
+        let foot = this.shape.children.find((c) => c.name === "ghostfoot");
+        if (foot) {foot.removeFromParent();}
+
+        let circle = new Microverse.THREE.CircleGeometry(0.3, 32);
+        circle.rotateX(-Math.PI / 2);
+        let material = new Microverse.THREE.MeshBasicMaterial({color: 0x666666, opacity: 0.3, transparent: true});
+        foot = new Microverse.THREE.Mesh(circle, material);
+        foot.position.set(0, -1.6, 0);
+        foot.name = "ghostfoot";
+
+        this.shape.add(foot);
+    }
+
     move(type, xyz) {
         if (!xyz) {return;}
-        this.say("poseAvatarRequest", {type, coordinates: xyz}, 30);
+        this.say("poseAvatarRequest", {type, coordinates: xyz, pointing: true}, 30);
     }
 
     avatarPosed(data) {
         if (!this.bones) {return;}
-        let {type, coordinates} = data;
-        if (type === "pointerMove") {
+
+        this.handedness = this.actor._cardData.handedness === "Left" ? "Left" : "Right";
+        this.otherHandName = this.actor._cardData.handedness === "Left" ? "RightHand" : "LeftHand";
+
+        let otherHand = this.bones.get(this.otherHandName);
+        otherHand.position.set();
+        let {type, coordinates, pointing} = data;
+        if (type === "pointerMove" || type === "move") {
             this.moveHead(coordinates);
         }
-        if (type === "keyDown" || type === "pointerDown" || type === "pointerUp" || type === "pointerTap") {
-            this.moveRightHand(coordinates);
+        let pointingChanged = this.isPointing !== pointing;
+        if (pointingChanged) {
+            this.isPointing = pointing;
+        }
+        if (type === "keyDown" || type === "pointerDown" || type === "pointerUp" || type === "pointerTap" || pointingChanged) {
+            this.moveHand(coordinates, pointing);
         }
     }
 
@@ -183,16 +192,18 @@ class AvatarPawn {
         rightEye.rotation.set(q_pitch(eyeQ), q_yaw(eyeQ), q_roll(eyeQ));
     }
 
-    moveRightHand(xyz) {
+    moveHand(xyz, pointing) {
         let {
             THREE,
             q_euler, q_pitch, q_yaw, q_roll, q_lookAt, q_multiply,// q_identity,
             v3_normalize, v3_rotate, v3_add} = Microverse;
-        let len = 0.6582642197608948 / 2;
-        let elbowPos = [-len, 0, 0.2];
-        let handPos = [0, 0, len * 0.6];
 
-        let hand = this.bones.get("RightHand");
+        let len = 0.6582642197608948 * 0.3;
+        let hFactor = this.handedness === "Left" ? -1 : 1;
+        let elbowPos = [-len * hFactor, 0, 0.2];
+        let handPos = [0, 0, hFactor * len * 0.3];
+
+        let hand = this.bones.get(`${this.handedness}Hand`);
         let spine = this.bones.get("Spine");
         let global = spine.matrixWorld.clone();
 
@@ -204,29 +215,30 @@ class AvatarPawn {
 
         global.multiply(dataRotation);
         global.multiply(elbowOffset);
-
         global.invert();
 
-        let local = new Microverse.THREE.Vector3(...xyz);
-        local.applyMatrix4(global);
-        let normLocal = v3_normalize(local.toArray());
-        let normHere = [0, 0, -1];
+        // console.log(this.actor._cardData.animationClipIndex);
+        if (!pointing) {
+            hand.rotation.set(Math.PI, hFactor * Math.PI / 2, 0);
+            hand.position.set();//...v3_add(elbowPos, [0, -0.25, -0.3]));
+            this.say("setAnimationClipIndex", 0);
+        } else {
+            this.say("setAnimationClipIndex", this.handedness === "Left" ? 12 : 9);
 
-        let allQ = q_lookAt(normHere, [0, 1, 0], normLocal);
+            let local = new Microverse.THREE.Vector3(...xyz);
+            local.applyMatrix4(global);
+            local = local.toArray();
+            local[2] = Math.min(local[2], -1);
+            let normLocal = v3_normalize(local);
+            let normHere = [0, 0, -1];
 
-        // console.log("hand", q_pitch(allQ), q_yaw(allQ), q_roll(allQ));
+            let allQ = q_lookAt(normHere, [0, 1, 0], normLocal);
+            // console.log("hand", q_pitch(allQ), q_yaw(allQ), q_roll(allQ));
+            let tQ = q_multiply(allQ, q_euler(-Math.PI / 2, 0, 0));
 
-        let tQ = q_euler(-q_pitch(allQ), q_roll(allQ), -q_yaw(allQ))
-
-        let handQ = q_multiply(q_euler(Math.PI / 2 - 0.2, 1.3, 0), tQ);
-        hand.rotation.set(q_pitch(handQ), q_yaw(handQ), q_roll(handQ));
-        hand.position.set(...v3_add(elbowPos, v3_rotate(handPos, tQ)));
-
-        let otherHand = this.bones.get("LeftHand");
-        let otherElbowPos = [len, 0, 0.2];
-        let otherQ = q_euler(Math.PI / 2, -0.4, 0.0);
-        otherHand.rotation.set(q_pitch(otherQ), q_yaw(otherQ), q_roll(otherQ));
-        otherHand.position.set(...otherElbowPos);
+            hand.rotation.set(-q_pitch(tQ), q_yaw(tQ), q_roll(tQ));
+            hand.position.set(...v3_add(elbowPos, v3_rotate(handPos, tQ)));
+        }
     }
 
     up(p3d) {
@@ -239,7 +251,7 @@ class AvatarPawn {
         let {m4_translation, q_axisAngle, m4_rotationQ, m4_multiply} = Microverse;
         const pitchRotation = q_axisAngle([1,0,0], this.lookPitch);
         const m0 = m4_translation(this.lookOffset);
-        const m1 = m4_translation([0, 0.4, 0]); // needs to be eye height;
+        const m1 = m4_translation([0, 0.2, 0]); // needs to be eye height;
         const tr = m4_multiply(m1, m0);
 
         const m2 = m4_rotationQ(pitchRotation);
@@ -247,12 +259,28 @@ class AvatarPawn {
         return m4_multiply(m3, this.global);
     }
 
-    mapOpacity(opacity) {
+    maybeMove() {
+        let velocity = Microverse.v3_magnitude(this.velocity);
+        let moving = velocity > 0.001;
+
+        let movingChanged = this.moving !== moving;
+
+        if (movingChanged) {
+            this.moving = moving;
+            let xyz = Microverse.v3_rotate([0, 0, -10], this.rotation);
+            this.say("poseAvatarRequest", {type: "move", coordinates: xyz, pointing: !this.moving}, 30);
+        }
+    }
+
+    mapOpacity(avatar, opacity) {
+        if (this._target === avatar && Microverse.v3_magnitude(this.lookOffset) < 0.8) {return 0;}
         if (opacity === 0 || opacity === 1) {return opacity;}
         return 1;
     }
 
     teardown() {
+        delete this.bones;
+        this.removeUpdateRequest(["HalfBodyAvatarEventHandler$AvatarPawn", "maybeMove"]);
         if (!this.isMyPlayerPawn) {return;}
         console.log("avatar event handler detached");
         this.removeFirstResponder("pointerTap", {ctrlKey: true, altKey: true}, this);