+++ behaviors/factory/rearrange.js
@@ -36,8 +36,7 @@ class RearrangeButtonPawn {
     }
 
     makeButton() {
-        this.shape.children.forEach((c) => this.shape.remove(c));
-        this.shape.children = [];
+        [...this.shape.children].forEach((c) => this.shape.remove(c));
 
         let geometry = new Microverse.THREE.SphereGeometry(0.15, 16, 16);
         let material = new Microverse.THREE.MeshStandardMaterial({color: 0xcccccc, metalness: 0.8});